[#19] Move parsing of the data into a separate ajax request
Motivation:

Rather than let the user put anything for the date, the code is better
if we specify that the params[:date] of an expense conforms to a defined
format.

However, it's nice to be able to type "2 jul" for the date - I know, I
do this a lot.  So lets parse that separately before the form is
submitted.